livekit-agent-server/server/requirements.txt
livekit==1.0.6
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
openai==1.2.0

# Note: The following packages may need to be installed from GitHub
# pip install git+https://github.com/livekit/agents-python.git
# pip install git+https://github.com/livekit/plugins-python.git
